+++ lib/open_book_web/controllers/session_controller.ex
@@ -5,7 +5,7 @@ defmodule OpenBookWeb.SessionController do
   alias OpenBook.LittleLogger, as: LL
   alias OpenBook.Plugs.Auth
 
-  alias OpenBookWeb.HomeLive
+  alias OpenBookWeb.ChallengesLive
 
   def login_through_url_with_verification_code(conn, %{"code" => code}) do
     LL.info_event("login_with_code", %{code: code})
@@ -23,7 +23,7 @@ defmodule OpenBookWeb.SessionController do
 
         conn
         |> Auth.login(user)
-        |> redirect(to: Routes.live_path(conn, HomeLive))
+        |> redirect(to: Routes.live_path(conn, ChallengesLive))
     end
   end
