+++ lib/open_book_web/live/markdowns/footer_menu_bar.ex
@@ -94,85 +94,84 @@ defmodule OpenBookWeb.Markdowns.FooterMenuBar do
   end
 end
 
-  ## Markdown
+## Markdown
 
-  # TODO(Arie): Delete.
-  # defp top_bar_nav(assigns) do
-  #   ~H"""
-  #   <div class="buttons top-bar has-addons is-centered pt-4 pb-4 mb-0 button-navbar">
-  #     <button
-  #       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "stats" })}
-  #       style="width: 110px"
-  #       phx-click="top_level_navigate"
-  #       phx-value-top_bar_tab="stats"
-  #     >
-  #       <span class="icon">
-  #         <i class="fas fa-trophy"></i>
-  #       </span>
-  #       <span>Stats</span>
-  #     </button>
+# TODO(Arie): Delete.
+# defp top_bar_nav(assigns) do
+#   ~H"""
+#   <div class="buttons top-bar has-addons is-centered pt-4 pb-4 mb-0 button-navbar">
+#     <button
+#       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "stats" })}
+#       style="width: 110px"
+#       phx-click="top_level_navigate"
+#       phx-value-top_bar_tab="stats"
+#     >
+#       <span class="icon">
+#         <i class="fas fa-trophy"></i>
+#       </span>
+#       <span>Stats</span>
+#     </button>
 
-  #     <button
-  #       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "book" })}
-  #       style="width: 110px"
-  #       phx-click="top_level_navigate"
-  #       phx-value-top_bar_tab="book"
-  #     >
-  #       <span class="icon">
-  #         <i class="fas fa-book"></i>
-  #       </span>
-  #       <span>Book</span>
-  #     </button>
+#     <button
+#       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "book" })}
+#       style="width: 110px"
+#       phx-click="top_level_navigate"
+#       phx-value-top_bar_tab="book"
+#     >
+#       <span class="icon">
+#         <i class="fas fa-book"></i>
+#       </span>
+#       <span>Book</span>
+#     </button>
 
-  #     <button
-  #       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "friends" })}
-  #       style="width: 110px"
-  #       phx-click="top_level_navigate"
-  #       phx-value-top_bar_tab="friends"
-  #     >
-  #       <span class="icon">
-  #         <i class="fas fa-user-friends"></i>
-  #       </span>
-  #       <span>Friends</span>
-  #     </button>
-  #   </div>
-  #   """
-  # end
+#     <button
+#       class={ViewUtils.class_list("button", %{"is-selected": @selected_tab == "friends" })}
+#       style="width: 110px"
+#       phx-click="top_level_navigate"
+#       phx-value-top_bar_tab="friends"
+#     >
+#       <span class="icon">
+#         <i class="fas fa-user-friends"></i>
+#       </span>
+#       <span>Friends</span>
+#     </button>
+#   </div>
+#   """
+# end
 
+# <.link
+#   navigate={Routes.live_path(OpenBookWeb.Endpoint, NutritionLogLive)}
+#   class="button is_quarter_width has-text-grey b-0 pb-4 m-0 br-0"
+# >
 
-  # <.link
-  #   navigate={Routes.live_path(OpenBookWeb.Endpoint, NutritionLogLive)}
-  #   class="button is_quarter_width has-text-grey b-0 pb-4 m-0 br-0"
-  # >
+#   <span>
+#     test
+#   </span>
+# </.link>
 
-  #   <span>
-  #     test
-  #   </span>
-  # </.link>
+# <.link
+#   navigate={Routes.live_path(OpenBookWeb.Endpoint, NutritionLogLive)}
+#   class="button is_quarter_width has-text-grey b-0 pb-4 m-0 br-0"
+# >
+#   <span class="icon">
+#     <i class="fas fa-book"></i>
+#   </span>
+# </.link>
 
-  # <.link
-  #   navigate={Routes.live_path(OpenBookWeb.Endpoint, NutritionLogLive)}
-  #   class="button is_quarter_width has-text-grey b-0 pb-4 m-0 br-0"
-  # >
-  #   <span class="icon">
-  #     <i class="fas fa-book"></i>
-  #   </span>
-  # </.link>
+# <.link
+#   navigate={Routes.live_path(OpenBookWeb.Endpoint, ExerciseLogLive)}
+#   class="button is_quarter_width has-text-grey is-white b-0 pb-4 m-0 br-0"
+# >
+#   <span class="icon">
+#     <i class="fas fa-user-friends"></i>
+#   </span>
+# </.link>
 
-  # <.link
-  #   navigate={Routes.live_path(OpenBookWeb.Endpoint, ExerciseLogLive)}
-  #   class="button is_quarter_width has-text-grey is-white b-0 pb-4 m-0 br-0"
-  # >
-  #   <span class="icon">
-  #     <i class="fas fa-user-friends"></i>
-  #   </span>
-  # </.link>
-
-  # <.link
-  #   navigate={Routes.live_path(OpenBookWeb.Endpoint, ExerciseLogLive)}
-  #   class="button is_quarter_width has-text-grey is-white b-0 pb-4 m-0 br-0"
-  # >
-  #   <span class="icon">
-  #     <i class="fas fa-pencil-alt"></i>
-  #   </span>
-  # </.link>
+# <.link
+#   navigate={Routes.live_path(OpenBookWeb.Endpoint, ExerciseLogLive)}
+#   class="button is_quarter_width has-text-grey is-white b-0 pb-4 m-0 br-0"
+# >
+#   <span class="icon">
+#     <i class="fas fa-pencil-alt"></i>
+#   </span>
+# </.link>
